movieStore | DeleteFix

+++ WebApi/Application/ActorOperations/Commands/DeleteActor/DeleteActorCommand.cs
@@ -17,6 +17,10 @@ namespace WebApi.Application.ActorOperations.Commands.DeleteActor
             var _actor = _context.Actors.SingleOrDefault(x => x.Id == actorID);
             if (_actor is null)
                 throw new InvalidOperationException("Actor bulunamadı.");
+            
+            var _movie = _context.Movies.FirstOrDefault(x => x.Actors.Any(u=> u.Id == actorID));
+            if (_movie is not null)
+                throw new InvalidOperationException("Actor silinemez. Mevcut Filmi Var.");
 
             _context.Actors.Remove(_actor);
             _context.SaveChanges();
